requirements.txt
aiofiles
aiohttp
matplotlib
numpy
opencv-python
pandas
Pillow
psutil
PyYAML
requests
scipy
seaborn
streamlit
supervision
thop
torch
torchvision
tqdm
ultralytics
Flask
flask_session
pycocotools
WTForms
roboflow